+++ test/apps/basics/src/routes/errors/invalid-route-response.js
@@ -1,0 +1,3 @@
+export function get() {
+  return 'this ought to be an object';
+}