+++ test/apps/options/snowpack.config.js
@@ -1,0 +1,10 @@
+// Consult https://www.snowpack.dev to learn about these options
+module.exports = {
+	extends: '@sveltejs/snowpack-config',
+	mount: {
+		'source/components': '/components'
+	},
+	alias: {
+		'$components': './source/components'
+	}
+};