+++ packages/adapter-node/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-node
 
+## 0.0.17
+
+### Patch Changes
+
+- Add svelte-kit start command
+
+## 0.0.16
+
+### Patch Changes
+
+- Make paths and target configurable
+
 ## 0.0.15
 
 ### Patch Changes