+++ packages/create-svelte/CHANGELOG.md
@@ -1,5 +1,18 @@
 # create-svelte
 
+## 2.0.0
+
+### Patch Changes
+
+- Add svelte-kit start command
+
+## 2.0.0-alpha.18
+
+### Patch Changes
+
+- rename CLI to svelte-kit
+- 0904e22: rename svelte CLI to svelte-kit
+
 ## 2.0.0-alpha.16
 
 ### Patch Changes